NodeConnector moved to common-jvm module
