+++ android/src/main/java/org/ergoplatform/android/settings/ConnectionSettingsDialogFragment.kt
@@ -6,6 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
+import org.ergoplatform.NodeConnector
 import org.ergoplatform.android.*
 import org.ergoplatform.android.databinding.FragmentConnectionSettingsBinding
