+++ android/src/main/java/org/ergoplatform/android/settings/DisplayCurrencyListDialogFragment.kt
@@ -13,7 +13,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
-import org.ergoplatform.android.NodeConnector
+import org.ergoplatform.NodeConnector
 import org.ergoplatform.android.Preferences
 import org.ergoplatform.android.R
 import org.ergoplatform.android.databinding.FragmentDisplayCurrencyDialogBinding
