+++ android/src/main/java/org/ergoplatform/android/transactions/ReceiveToWalletFragment.kt
@@ -10,20 +10,17 @@ import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.navArgs
 import kotlinx.coroutines.launch
 import org.ergoplatform.android.AppDatabase
-import org.ergoplatform.android.NodeConnector
+import org.ergoplatform.NodeConnector
 import org.ergoplatform.android.R
 import org.ergoplatform.android.databinding.FragmentReceiveToWalletBinding
-import org.ergoplatform.android.ui.copyStringToClipboard
-import org.ergoplatform.android.ui.formatFiatToString
-import org.ergoplatform.android.ui.inputTextToDouble
-import org.ergoplatform.android.ui.setQrCodeToImageView
+import org.ergoplatform.android.ui.*
 import org.ergoplatform.android.wallet.WalletDbEntity
 import org.ergoplatform.android.wallet.addresses.AddressChooserCallback
 import org.ergoplatform.android.wallet.addresses.ChooseAddressListDialogFragment
-import org.ergoplatform.android.wallet.addresses.getAddressLabel
 import org.ergoplatform.android.wallet.getDerivedAddress
 import org.ergoplatform.android.wallet.getDerivedAddressEntity
 import org.ergoplatform.getExplorerPaymentRequestAddress
+import org.ergoplatform.wallet.addresses.getAddressLabel
 
 
 /**
@@ -92,7 +89,7 @@ class ReceiveToWalletFragment : Fragment(), AddressChooserCallback {
 
     private fun refreshAddressInformation() {
         val address = wallet?.getDerivedAddressEntity(derivationIdx)
-        binding.addressLabel.text = address?.getAddressLabel(requireContext())
+        binding.addressLabel.text = address?.getAddressLabel(AndroidStringProvider(requireContext()))
         binding.publicAddress.text = address?.publicAddress
 
         refreshQrCode()
